+++ packages/falcon-wordpress-api/src/index.js
@@ -6,6 +6,14 @@ const isObject = require('lodash/isObject');
 const url = require('url');
 
 module.exports = class WordpressApi extends ApiDataSource {
+  constructor(params) {
+    super(params);
+    this.wpConfig = {};
+    this.languageMap = {};
+    this.languageSupported = false;
+    this.baseLanguage = null;
+  }
+
   async authorizeRequest(req) {
     const { username, password } = this.config;
     const token = Buffer.from(`${username}:${password}`).toString('base64');
@@ -18,18 +26,16 @@ module.exports = class WordpressApi extends ApiDataSource {
    * @return {string} resolved url object
    */
   async resolveURL(req) {
-    const { path, context = {} } = req;
-    const { language } = context;
-    const { apiPrefix, apiSuffix, language: baseLanguage } = this.config;
-    let prefix = `${apiPrefix}${apiSuffix}`;
+    const { path } = req;
+    const { locale } = this.context.session || {};
+    const { apiPrefix } = this.config;
+    const { baseLanguage, languageSupported, languageMap } = this;
 
-    if (language) {
-      // for base lang do not add prefix
-      const languagePrefix = language && language !== baseLanguage ? `/${language}` : '';
-      prefix = `${languagePrefix}${prefix}`;
-    }
+    const language = languageSupported && locale && languageMap[locale] ? languageMap[locale] : baseLanguage;
+    // for base lang do not add prefix
+    const langPrefix = language && language !== baseLanguage ? `/${language}` : '';
 
-    return super.resolveURL({ path: `${prefix}/${path}` });
+    return super.resolveURL({ path: `${langPrefix}${apiPrefix}/${path}` });
   }
 
   /**
@@ -59,24 +65,35 @@ module.exports = class WordpressApi extends ApiDataSource {
     };
   }
 
-  async preInitialize() {
-    if (!this.context) {
-      this.initialize({ context: {} });
+  async fetchBackendConfig() {
+    const config = await this.get('blog/info');
+    const { languages = {} } = config;
+    this.languageSupported = !!Object.keys(languages).length;
+
+    if (this.languageSupported) {
+      this.wpConfig.locales = [];
+      this.baseLanguage = languages.default;
+      languages.options.forEach(({ default_locale: locale, language_code: langCode }) => {
+        this.languageMap[locale] = langCode;
+        this.wpConfig.locales.push(locale);
+        if (this.baseLanguage === langCode) {
+          this.wpConfig.defaultLocale = locale;
+        }
+      });
     }
-    return this.get('blog/info');
+
+    return this.wpConfig;
   }
 
   /**
    * Fetch single published post by slug
    * @query
-   * @param {object} root GraphQL root object
+   * @param {object} _ GraphQL root object
    * @param {string} path WP "slug" value
-   * @param {object} session Web-server session object
    * @return {Object} Post data
    */
-  async blogPost(root, { path }, { session }) {
+  async blogPost(_, { path }) {
     const slug = path.replace('/', '');
-    const { language } = session;
 
     const query = {
       slug,
@@ -87,7 +104,6 @@ module.exports = class WordpressApi extends ApiDataSource {
 
     return this.get('posts', query, {
       context: {
-        language,
         didReceiveResult: (result, res) => {
           // WP API returns "post" entry as an array, the following code removes extra-headers
           if (res && res.headers && res.headers.has('x-wp-total')) {
@@ -103,14 +119,12 @@ module.exports = class WordpressApi extends ApiDataSource {
   /**
    * Fetch published posts.
    * @query
-   * @param {object} root GraphQL root object
+   * @param {object} _ GraphQL root object
    * @param {object} query Query object
-   * @param {object} session Web-server session data
+   * @param {object} pagination Pagination
    * @return {Object[]} posts data
    */
-  async blogPosts(root, { query, pagination }, { session }) {
-    const { language } = session;
-
+  async blogPosts(_, { query, pagination }) {
     const payload = {
       ...query
     };
@@ -122,7 +136,6 @@ module.exports = class WordpressApi extends ApiDataSource {
 
     return this.get('posts', payload, {
       context: {
-        language,
         didReceiveResult: result =>
           result && Array.isArray(result) ? result.map(entry => this.processPost(entry)) : result
       }
@@ -339,23 +352,24 @@ module.exports = class WordpressApi extends ApiDataSource {
   /**
    * Fetch wordpress url based on pathname and check if it contains any redirect.
    * Convert response based on data type (page | post | category )
-   * @param {String} path - pathname of wordpress url
-   * @param {String} language - wordpress language
+   * @param {object} _ GQL root object
+   * @param {object} params GQL params object
+   * @param {string} params.path URL path param
    * @return {Object} response - with reduced and converted data
    */
-  async fetchUrl(path, language = null) {
-    const params = { path };
-    if (language) {
-      params.language = language;
-    }
+  async fetchUrl(_, { path }) {
+    const { locale } = this.context.session;
 
-    return this.get('url', params, {
-      context: {
-        authRequired: this.isDraft(path),
-        language,
-        didReceiveResult: result => this.reduceUrl(result, path, language)
+    return this.get(
+      'url',
+      { path },
+      {
+        context: {
+          authRequired: this.isDraft(path),
+          didReceiveResult: result => this.reduceUrl(result, path, this.languageMap[locale])
+        }
       }
-    });
+    );
   }
 
   reduceUrl(result, path, language) {