+++ CHANGELOG.md
@@ -4,10 +4,20 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ## Falcon v0.2
 
-### Falcon Client v0.0.3
+### Falcon Client v0.0.8
 
 - feat: removed razzle ([#87](https://github.com/deity-io/falcon/pull/87))
 - feat: support for Google Analytics added ([#78](https://github.com/deity-io/falcon/pull/78))
+
+### Falcon Server v0.0.4
+
+- feat: Basic Cache implementation was introduced ([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: (Breaking change) changed Event flow for ApiContainer and its entries - every ApiDataSource instance is being
+created on GQL request ([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: added `backendConfig` Query type ([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: `type BackendConfig`, `enum SortOrderDirection`, `input SortOrder` were introduced in the base Schema ([#176](https://github.com/deity-io/falcon/pull/176))
+- fix: added separate `endpoints` config section and dedicated base class ([#176](https://github.com/deity-io/falcon/pull/176))
+- refactor: `Events` enum has been moved from `falcon-server` to `falcon-server-env` package ([#176](https://github.com/deity-io/falcon/pull/176))
 
 ### Falcon UI v0.0.4
 
@@ -31,6 +41,21 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 - feat: allow ejecting `falcon-ecommerce-uikit` package via `eject` command (https://github.com/deity-io/falcon/pull/212))
 - docs: updated documentation ([#47](https://github.com/deity-io/falcon/pull/47))
 - fix: fixed problem with React 16.6.0 ([#109](https://github.com/deity-io/falcon/pull/109))
+
+### Falcon Server Env v0.0.4
+
+- feat: Provided `Cache` wrapper class and built-in `InMemoryLRUCache` cache provider ([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: added getter and setter methods to work with "named" session object from the context in ApiDataSource
+([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: `ApiDataSource` and `Extension` models are now accept `eventEmitter` instance
+([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: Provided `EndpointManager` base class ([#176](https://github.com/deity-io/falcon/pull/176))
+- feat: Base `Extension` class provides auto-binding for its own GraphQL Schema to the assigned ApiDataSource instance
+(via `getGraphQLConfig()` method) ([#176](https://github.com/deity-io/falcon/pull/176))
+
+### Falcon Magento2 API v0.0.4
+
+- feat: Magento Admin token is now being stored in cache ([#176](https://github.com/deity-io/falcon/pull/176))
 
 ### eject-ts v0.0.1